what changed: build fixed how tested: run

+++ _studio/mfx_lib/audio_encode/aac/include/mfx_aac_encode.h
@@ -94,7 +94,7 @@ protected:
         CommonCORE& mCore;
         std::list<mfxAudioFrame*> list;
         std::vector<mfxU8>& buffer;
-        int offset;
+        size_t offset;
     public:
         AudioFramesCollector(std::vector<mfxU8>& v, CommonCORE& core)
             : offset(0)
